Fix bug when building non-switch and appending channel hash
When we build a non-switch (e.g. "boot") and append the channel hash in
the config tag, we append the wrong hash (not the hash of the channel of
the build generation but the hash of the current channel).

This is a severe bug which is fixed with this commit (by simply failing
to allow the combination of these two flags).